+++ src/main/java/com/nhl/bootique/jersey/client/auth/AuthenticatorFactory.java
@@ -1,6 +1,7 @@
 package com.nhl.bootique.jersey.client.auth;
 
 import javax.ws.rs.client.ClientRequestFilter;
+import javax.ws.rs.core.Configuration;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.nhl.bootique.config.PolymorphicConfiguration;
@@ -9,7 +10,7 @@ import com.nhl.bootique.config.PolymorphicConfiguration;
  * @since 0.2
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-public interface AuthenticatorFactory extends PolymorphicConfiguration{
+public interface AuthenticatorFactory extends PolymorphicConfiguration {
 
-	ClientRequestFilter createAuthFilter();
+	ClientRequestFilter createAuthFilter(Configuration filterFreeConfig);
 }